Clean up IndexState
