+++ src/main/java/com/yelp/nrtsearch/server/luceneserver/RecvCopyStateHandler.java
@@ -30,7 +30,7 @@ public class RecvCopyStateHandler implements Handler<CopyStateRequest, CopyState
     ShardState shardState = indexState.getShard(0);
     if (shardState.isPrimary() == false) {
       throw new IllegalArgumentException(
-          "index \"" + indexState.name + "\" was not started or is not a primary");
+          "index \"" + indexState.getName() + "\" was not started or is not a primary");
     }
 
     if (!isValidMagicHeader(copyStateRequest.getMagicNumber())) {
