+++ src/test/java/com/yelp/nrtsearch/server/grpc/TimeoutTest.java
@@ -484,7 +484,7 @@ public class TimeoutTest extends ServerTestCase {
       DrillSideways drillS =
           new DrillSidewaysImpl(
               context.getSearcherAndTaxonomy().searcher,
-              context.getIndexState().facetsConfig,
+              context.getIndexState().getFacetsConfig(),
               context.getSearcherAndTaxonomy().taxonomyReader,
               Collections.singletonList(getTestFacet()),
               context.getSearcherAndTaxonomy(),
